@@ -1,11 +1,10 @@
 from __future__ import annotations
 
-import warnings
 from abc import ABC, abstractmethod
 from collections.abc import MutableMapping, Sequence
 from copy import copy
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Generic, TypeVar
+from typing import TYPE_CHECKING
 
 import numpy as np
 import pandas as pd
@@ -16,7 +15,9 @@ from ..utils import (
     axis_len,
     convert_to_dict,
     deprecated,
+    deprecation_msg,
     raise_value_error_if_multiindex_columns,
+    warn,
     warn_once,
 )
 from .access import ElementRef
@@ -38,12 +39,8 @@ TwoDIdx = tuple[OneDIdx, OneDIdx]
 # TODO: pd.DataFrame only allowed in AxisArrays?
 Value = pd.DataFrame | CSMatrix | CSArray | np.ndarray
 
-P = TypeVar("P", bound="AlignedMappingBase")
-"""Parent mapping an AlignedView is based on."""
-I = TypeVar("I", OneDIdx, TwoDIdx)
 
-
-class AlignedMappingBase(MutableMapping[str, Value], ABC):
+class AlignedMappingBase[I: OneDIdx](MutableMapping[str, Value], ABC):
     """\
     An abstract base class for Mappings containing array-like values aligned
     to either one or both AnnData axes.
@@ -70,12 +67,11 @@ class AlignedMappingBase(MutableMapping[str, Value], ABC):
     def _validate_value(self, val: Value, key: str) -> Value:
         """Raises an error if value is invalid"""
         if isinstance(val, AwkArray):
-            warn_once(
+            msg = (
                 "Support for Awkward Arrays is currently experimental. "
-                "Behavior may change in the future. Please report any issues you may encounter!",
-                ExperimentalFeatureWarning,
-                # stacklevel=3,
+                "Behavior may change in the future. Please report any issues you may encounter!"
             )
+            warn_once(msg, ExperimentalFeatureWarning)
         elif isinstance(val, np.ndarray | CupyArray) and len(val.shape) == 1:
             val = val.reshape((val.shape[0], 1))
         elif isinstance(val, XDataset):
@@ -130,12 +126,12 @@ class AlignedMappingBase(MutableMapping[str, Value], ABC):
         """Returns a subset copy-on-write view of the object."""
         return self._view_class(self, parent, subset_idx)
 
-    @deprecated("dict(obj)")
+    @deprecated(deprecation_msg("as_dict", "dict(obj)"))
     def as_dict(self) -> dict:
         return dict(self)
 
 
-class AlignedView(AlignedMappingBase, Generic[P, I]):
+class AlignedView[P: AlignedMappingBase, I: (OneDIdx, TwoDIdx)](AlignedMappingBase):
     is_view: ClassVar[Literal[True]] = True
 
     # override docstring
@@ -151,7 +147,7 @@ class AlignedView(AlignedMappingBase, Generic[P, I]):
     subset_idx: I
     """The subset of the parent to view."""
 
-    def __init__(self, parent_mapping: P, parent_view: AnnData, subset_idx: I):
+    def __init__(self, parent_mapping: P, parent_view: AnnData, subset_idx: I) -> None:
         self.parent_mapping = parent_mapping
         self._parent = parent_view
         self.subset_idx = subset_idx
@@ -167,12 +163,11 @@ class AlignedView(AlignedMappingBase, Generic[P, I]):
 
     def __setitem__(self, key: str, value: Value) -> None:
         value = self._validate_value(value, key)  # Validate before mutating
-        warnings.warn(
+        msg = (
             f"Setting element `.{self.attrname}['{key}']` of view, "
-            "initializing view as actual.",
-            ImplicitModificationWarning,
-            stacklevel=2,
+            "initializing view as actual."
         )
+        warn(msg, ImplicitModificationWarning)
         with view_update(self.parent, self.attrname, ()) as new_mapping:
             new_mapping[key] = value
 
@@ -180,12 +175,11 @@ class AlignedView(AlignedMappingBase, Generic[P, I]):
         if key not in self:
             msg = f"{key!r} not found in view of {self.attrname}"
             raise KeyError(msg)  # Make sure it exists before bothering with a copy
-        warnings.warn(
+        msg = (
             f"Removing element `.{self.attrname}['{key}']` of view, "
-            "initializing view as actual.",
-            ImplicitModificationWarning,
-            stacklevel=2,
+            "initializing view as actual."
         )
+        warn(msg, ImplicitModificationWarning)
         with view_update(self.parent, self.attrname, ()) as new_mapping:
             del new_mapping[key]
 
@@ -381,7 +375,7 @@ PairwiseArraysBase._view_class = PairwiseArraysView
 PairwiseArraysBase._actual_class = PairwiseArrays
 
 
-AlignedMapping = (
+type AlignedMapping = (
     AxisArrays
     | AxisArraysView
     | Layers
@@ -389,12 +383,11 @@ AlignedMapping = (
     | PairwiseArrays
     | PairwiseArraysView
 )
-T = TypeVar("T", bound=AlignedMapping)
 """Pair of types to be aligned."""
 
 
 @dataclass
-class AlignedMappingProperty(property, Generic[T]):
+class AlignedMappingProperty[T: AlignedMapping](property):
     """A :class:`property` that creates an ephemeral AlignedMapping.
 
     The actual data is stored as `f'_{self.name}'` in the parent object.
