@@ -19,7 +19,7 @@ from ._warnings import (
     WriteWarning,
 )
 from .io import read_h5ad, read_zarr
-from .utils import module_get_attr_redirect
+from .utils import module_get_attr_redirect, warn
 
 # Submodules need to be imported last
 from . import abc, experimental, typing, io, types  # isort: skip
@@ -60,11 +60,10 @@ _DEPRECATED = {method: f"io.{method}" for method in _DEPRECATED_IO}
 
 def __getattr__(attr_name: str) -> Any:
     if attr_name == "__version__":
-        import warnings
         from importlib.metadata import version
 
         msg = "`__version__` is deprecated, use `importlib.metadata.version('anndata')` instead."
-        warnings.warn(msg, FutureWarning, stacklevel=2)
+        warn(msg, FutureWarning)
         return version("anndata")
 
     return module_get_attr_redirect(attr_name, deprecated_mapping=_DEPRECATED)
