@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-import warnings
+from ..utils import warn
 
 __all__: list[str] = []
 
@@ -13,5 +13,5 @@ def __getattr__(key: str):
         f"Importing {key} from `anndata._io` is deprecated. "
         "Please use anndata.io instead."
     )
-    warnings.warn(msg, FutureWarning, stacklevel=2)
+    warn(msg, FutureWarning)
     return attr
