@@ -1,8 +1,9 @@
 from __future__ import annotations
 
 import re
+import sys
 import warnings
-from functools import singledispatch, wraps
+from functools import partial, singledispatch
 from typing import TYPE_CHECKING
 
 import h5py
@@ -13,12 +14,13 @@ from scipy import sparse
 import anndata
 
 from ._core.sparse_dataset import BaseCompressedSparseDataset
+from ._warnings import warn
 from .compat import CSArray, CupyArray, CupySparseMatrix, DaskArray
 from .logging import get_logger
 
 if TYPE_CHECKING:
     from collections.abc import Iterable, Mapping, Sequence
-    from typing import Any, Literal
+    from typing import Any, Literal, LiteralString
 
 logger = get_logger(__name__)
 
@@ -272,8 +274,7 @@ def make_index_unique(index: pd.Index, join: str = "-"):
             "Example key collisions generated by the make_index_unique algorithm: "
             f"{example_colliding_values}"
         )
-        # 3: caller -> 2: `{obs,var}_names_make_unique` -> 1: here
-        warnings.warn(msg, UserWarning, stacklevel=3)
+        warn(msg, UserWarning)
     values[indices_dup] = values_dup
     index = pd.Index(values, name=index.name)
     return index
@@ -292,12 +293,11 @@ def join_english(words: Iterable[str], conjunction: str = "or") -> str:
 
 def warn_names_duplicates(attr: str):
     names = "Observation" if attr == "obs" else "Variable"
-    warnings.warn(
+    msg = (
         f"{names} names are not unique. "
-        f"To make them unique, call `.{attr}_names_make_unique`.",
-        UserWarning,
-        stacklevel=2,
+        f"To make them unique, call `.{attr}_names_make_unique`."
     )
+    warn(msg, UserWarning)
 
 
 def ensure_df_homogeneous(
@@ -310,8 +310,7 @@ def ensure_df_homogeneous(
         arr = df.to_numpy()
     if df.dtypes.nunique() != 1:
         msg = f"{name} converted to numpy array with dtype {arr.dtype}"
-        # 4: caller -> 3: `AnnData.__init__` -> 2: `_init_as_actual` → 1: here
-        warnings.warn(msg, UserWarning, stacklevel=4)
+        warn(msg, UserWarning)
     return arr
 
 
@@ -358,44 +357,33 @@ def convert_dictionary_to_structured_array(source: Mapping[str, Sequence[Any]]):
     return arr
 
 
-def warn_once(msg: str, category: type[Warning], stacklevel: int = 1):
-    warnings.warn(msg, category, stacklevel=stacklevel)
+def warn_once(msg: str, category: type[Warning]) -> None:
+    warn(msg, category)
     # Prevent from showing up every time an awkward array is used
     # You'd think `'once'` works, but it doesn't at the repl and in notebooks
     warnings.filterwarnings("ignore", category=category, message=re.escape(msg))
 
 
-def deprecated(
-    new_name: str,
-    category: type[Warning] = FutureWarning,
-    add_msg: str = "",
-    *,
-    hide: bool = True,
-):
-    """\
-    This is a decorator which can be used to mark functions
-    as deprecated with a FutureWarning. It will result in a warning being emitted
-    when the function is used.
-    """
+if TYPE_CHECKING:
+    from warnings import deprecated
+else:
+    if sys.version_info >= (3, 13):
+        from warnings import deprecated as _deprecated
+    else:
+        from typing_extensions import deprecated as _deprecated
+    deprecated = partial(_deprecated, category=FutureWarning)
 
-    def decorator(func):
-        name = func.__qualname__
-        msg = (
-            f"Use {new_name} instead of {name}, "
-            f"{name} is deprecated and will be removed in the future."
-        )
-        if add_msg:
-            msg += f" {add_msg}"
 
-        @wraps(func)
-        def new_func(*args, **kwargs):
-            warnings.warn(msg, category=category, stacklevel=2)
-            return func(*args, **kwargs)
-
-        setattr(new_func, "__deprecated", (category, msg, hide))
-        return new_func
-
-    return decorator
+def deprecation_msg(
+    name: LiteralString, new_name: LiteralString, add_msg: LiteralString | None = None
+) -> LiteralString:
+    msg = (
+        f"Use {new_name} instead of {name}, "
+        f"{name} is deprecated and will be removed in the future."
+    )
+    if add_msg is not None:
+        msg += f" {add_msg}"
+    return msg
 
 
 class DeprecationMixinMeta(type):
@@ -405,11 +393,13 @@ class DeprecationMixinMeta(type):
     """
 
     def __dir__(cls):
-        def is_hidden(attr) -> bool:
+        dont_hide = getattr(cls, "_DONT_HIDE_DEPRECATED", set())
+
+        def is_hidden(attr: object) -> bool:
             if isinstance(attr, property):
                 attr = attr.fget
-            _, _, hide = getattr(attr, "__deprecated", (None, None, False))
-            return hide
+            is_deprecated = bool(getattr(attr, "__deprecated__", None))
+            return is_deprecated and getattr(attr, "__name__", None) not in dont_hide
 
         return [
             item
@@ -440,7 +430,7 @@ def module_get_attr_redirect(
             f"Importing {attr_name} from `{full_old_module_path}` is deprecated. "
             f"Import anndata.{new_path} instead."
         )
-        warnings.warn(msg, FutureWarning, stacklevel=2)
+        warn(msg, FutureWarning)
         # hacky import_object_by_name, but we test all these
         mod = anndata
         while "." in new_path:
