@@ -11,6 +11,7 @@ import zarr
 from anndata import AnnData
 from anndata.compat import DaskArray
 from anndata.experimental import read_elem_lazy, read_lazy
+from anndata.experimental.backed._io import ANNDATA_ELEMS
 from anndata.io import write_elem
 from anndata.tests.helpers import (
     GEN_ADATA_NO_XARRAY_ARGS,
@@ -19,8 +20,6 @@ from anndata.tests.helpers import (
     gen_adata,
     gen_typed_df,
 )
-
-from .conftest import ANNDATA_ELEMS
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -175,7 +174,7 @@ def test_view_of_view_to_memory(adata_remote: AnnData, adata_orig: AnnData):
 
 @pytest.mark.zarr_io
 def test_unconsolidated(tmp_path: Path, mtx_format):
-    adata = gen_adata((1000, 1000), mtx_format, **GEN_ADATA_NO_XARRAY_ARGS)
+    adata = gen_adata((10, 10), mtx_format, **GEN_ADATA_NO_XARRAY_ARGS)
     orig_pth = tmp_path / "orig.zarr"
     adata.write_zarr(orig_pth)
     (orig_pth / ".zmetadata").unlink()
@@ -186,6 +185,16 @@ def test_unconsolidated(tmp_path: Path, mtx_format):
     remote_to_memory = remote.to_memory()
     assert_equal(remote_to_memory, adata)
     store.assert_access_count("obs/.zgroup", 1)
+
+
+def test_h5_file_obj(tmp_path: Path):
+    adata = gen_adata((10, 10), **GEN_ADATA_NO_XARRAY_ARGS)
+    orig_pth = tmp_path / "adata.h5ad"
+    adata.write_h5ad(orig_pth)
+    remote = read_lazy(orig_pth)
+    assert remote.file.is_open
+    assert remote.filename == orig_pth
+    assert_equal(remote.to_memory(), adata)
 
 
 @pytest.fixture(scope="session")
